requirements.txt
asgiref==3.8.1
certifi==2024.8.30       
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3     
defusedxml==0.8.0rc2
Django==5.1.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
django-filter==23.1